test score

+++ transformer.py
@@ -340,7 +340,7 @@ class Transformer:
 		compressed_size = 0
 		tile_size = tilew * tileh
 		for roi,r in zip(ROIs,quality):
-			dsize = (np.rint(tilew*r),np.rint(tileh*r))
+			dsize = (int(np.rint(tilew*r)),int(np.rint(tileh*r)))
 			print(roi,dsize)
 			if dsize == (tilew,tileh):
 				compressed_size += tilew*tileh
